// packages/plugin-issue/src/generated-admin-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string | number;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
  Money: any;
  Upload: any;
};
// ...
export type CampaignFilterParameter = {
  campaignType?: InputMaybe<StringOperators>;
  code?: InputMaybe<StringOperators>;
  createdAt?: InputMaybe<DateOperators>;
  enabled?: InputMaybe<BooleanOperators>;
  id?: InputMaybe<IdOperators>;
  lgImage?: InputMaybe<StringOperators>;
  name?: InputMaybe<StringOperators>;
  needClaimCoupon?: InputMaybe<BooleanOperators>;
  periodOfValidity?: InputMaybe<NumberOperators>;
  shortDesc?: InputMaybe<StringOperators>;
  updatedAt?: InputMaybe<DateOperators>;
  xsImage?: InputMaybe<StringOperators>;
};

export type CampaignList = PaginatedList & {
  __typename?: 'CampaignList';
  items: Array<Campaign>;
  totalItems: Scalars['Int'];
};

export type CampaignListOptions = {
  /** Allows the results to be filtered */
  filter?: InputMaybe<CampaignFilterParameter>;
  /** Specifies whether multiple "filter" arguments should be combines with a logical AND or OR operation. Defaults to AND. */
  filterOperator?: InputMaybe<LogicalOperator>;
  /** Skips the first n results, for use in pagination */
  skip?: InputMaybe<Scalars['Int']>;
  /** Specifies which properties to sort the results by */
  sort?: InputMaybe<CampaignSortParameter>;
  /** Takes n results, for use in pagination */
  take?: InputMaybe<Scalars['Int']>;
};

export type CampaignSortParameter = {
  code?: InputMaybe<SortOrder>;
  createdAt?: InputMaybe<SortOrder>;
  id?: InputMaybe<SortOrder>;
  lgImage?: InputMaybe<SortOrder>;
  name?: InputMaybe<SortOrder>;
  periodOfValidity?: InputMaybe<SortOrder>;
  shortDesc?: InputMaybe<SortOrder>;
  updatedAt?: InputMaybe<SortOrder>;
  xsImage?: InputMaybe<SortOrder>;
};

export type CampaignTranslationInput = {
  applyForBeforeData: CampaignApplyForBeforeDataInput;
  applyForResultData?: InputMaybe<CampaignApplyForResultDataInput>;
  id?: InputMaybe<Scalars['ID']>;
  languageCode: LanguageCode;
  lgImage?: InputMaybe<Scalars['String']>;
  name: Scalars['String'];
  shortDesc?: InputMaybe<Scalars['String']>;
  xsImage?: InputMaybe<Scalars['String']>;
};

export enum CampaignType {
  /** 所有用户, 固定金额拉新, 直降模式 */
  AllUserFixedAmount = 'AllUserFixedAmount',
  /** 买一送一+镜片打折 */
  BuyXGetYFreeWithLensDiscount = 'BuyXGetYFreeWithLensDiscount',
  /** 节假日直活动 */
  DirectDiscount = 'DirectDiscount',
  /** 新用户首单镜架1USD */
  NewUserFirstPair = 'NewUserFirstPair',
  /** 新用户固定金额拉新, 领券模式 */
  NewUserFixedAmount = 'NewUserFixedAmount',
}
// ...
export type CreateCampaignInput = {
  campaignType: CampaignType;
  code: Scalars['String'];
  enabled: Scalars['Boolean'];
  needClaimCoupon?: InputMaybe<Scalars['Boolean']>;
  periodOfValidity?: InputMaybe<Scalars['Int']>;
  promotionId?: InputMaybe<Scalars['Int']>;
  translations: Array<CampaignTranslationInput>;
};
// ...
export type DeletionResponse = {
  __typename?: 'DeletionResponse';
  message?: Maybe<Scalars['String']>;
  result: DeletionResult;
};

export enum DeletionResult {
  /** The entity was successfully deleted */
  DELETED = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NOT_DELETED = 'NOT_DELETED',
}
// ...
export enum LogicalOperator {
  AND = 'AND',
  OR = 'OR',
}
// ...
export type MutationCreateCampaignArgs = {
  input: CreateCampaignInput;
};
// ...
export type MutationDeleteCampaignArgs = {
  id: Scalars['ID'];
};
// ...
export type MutationUpdateCampaignArgs = {
  input: UpdateCampaignInput;
};
// ...
export type QueryCampaignArgs = {
  id: Scalars['ID'];
};

export type QueryCampaignCouponArgs = {
  id: Scalars['ID'];
};

export type QueryCampaignCouponsArgs = {
  options?: InputMaybe<CampaignCouponListOptions>;
};

export type QueryCampaignsArgs = {
  options?: InputMaybe<CampaignListOptions>;
};
// ...
export type UpdateCampaignInput = {
  campaignType?: InputMaybe<CampaignType>;
  code?: InputMaybe<Scalars['String']>;
  enabled?: InputMaybe<Scalars['Boolean']>;
  id: Scalars['ID'];
  needClaimCoupon?: InputMaybe<Scalars['Boolean']>;
  periodOfValidity?: InputMaybe<Scalars['Int']>;
  promotionId?: InputMaybe<Scalars['Int']>;
  translations?: InputMaybe<Array<CampaignTranslationInput>>;
};